% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MTAFT_test.R
\name{MTAFT_test}
\alias{MTAFT_test}
\title{Perform score-type test for the presence of threshold effect in multi-threshold situations.}
\usage{
MTAFT_test(Y, X, Tq, delta, nboots)
}
\arguments{
\item{Y}{Response variable.}

\item{X}{Covariates.}

\item{Tq}{Threshold variable.}

\item{delta}{Indicator vector for censoring.}

\item{nboots}{Number of bootstrap iterations.}
}
\value{
p-value result indicating the presence of threshold effect.
}
\description{
This function performs a score-type test statistics for the presence of threshold effect in multi-threshold situations.
}
\examples{
# Perform score-type test with 500 bootstraps
pval <- mAFT_test(Y, X, Tq, delta, nboots = 500)

# Perform score-type test with 1000 bootstraps
pval <- MTAFT_test(Y, X, Tq, delta, nboots = 1000)
}
